requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
sqlalchemy>=1.4.0,<1.5.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
celery>=5.2.3,<5.3.0
python-dotenv>=0.19.0,<0.20.0
psycopg2-binary>=2.9.1,<2.10.0
aiohttp>=3.8.1,<3.9.0
beautifulsoup4>=4.9.3,<4.10.0
python-magic>=0.4.24,<0.5.0
Pillow>=8.3.2,<8.4.0
requests>=2.26.0,<2.27.0
lxml>=4.9.0,<4.10.0
cryptography>=41.0.0,<42.0.0 